+++ srlc/srl/templatetags/custom_filters.py
@@ -1,5 +1,6 @@
 from django import template
 from django.db.models import Sum
+from django.utils.timezone import now
 from ..models import GameOverview,ILRuns,MainRuns,Players,CountryCodes
 
 register = template.Library()
@@ -32,16 +33,6 @@ def filter_game_name(game_runs, game_name):
     return [game for game in game_runs if GameOverview.objects.get(id=game.game.id).name == game_name]
 
 @register.filter
-def get_game_boxart(games, game_id):
-    game = games.get(id=game_id)
-    return game.boxart
-
-@register.filter
-def is_il_cat(game_name):
-    thps_games = ["Tony Hawk's Pro Skater", "Tony Hawk's Pro Skater 2", "Tony Hawk's Pro Skater 3", "Tony Hawk's Pro Skater 4", "Tony Hawk's Pro Skater 1+2"]
-    return game_name in thps_games
-
-@register.filter
 def get_rank(game_name, player_name):
     players = Players.objects.all()
     leaderboard = []
@@ -65,12 +56,25 @@ def get_rank(game_name, player_name):
     for entry in il_leaderboard:
         if entry["player"] == player_name:
             return entry
-
+        
 @register.filter
-def get_item(dictionary, key):
-    return dictionary.get(key)
+def time_since(value):
+    if not value:
+        return ""
 
-@register.filter
-def get_country(countrycode):
-    country = CountryCodes.objects.filter(id=countrycode).values("name")[0]["name"]
-    return country
+    delta = now() - value
+
+    hours = delta.seconds // 3600
+    minutes = (delta.seconds % 3600) // 60
+
+    if delta.days > 0:
+        hours += delta.days * 24
+
+    if hours and minutes:
+        return f"Started {hours} hours and {minutes} minutes ago"
+    elif hours:
+        return f"Started {hours} hours ago"
+    elif minutes:
+        return f"Started {minutes} minutes ago"
+    else:
+        return "Just now"