+++ srlc/srl/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .complex_views import MainPage,PlayerProfile,Leaderboard,GameLeaderboard,IL_Leaderboard,FG_Leaderboard,search_leaderboard,ILGameLeaderboard,MonthlyLeaderboard
+from .complex_views import MainPage,PlayerProfile,PlayerHistory,Leaderboard,GameLeaderboard,IL_Leaderboard,FG_Leaderboard,search_leaderboard,ILGameLeaderboard,MonthlyLeaderboard
 from .static_views import PrivacyPolicy,Changelog,FAQ
 
 urlpatterns = [
@@ -12,6 +12,7 @@ urlpatterns = [
     #path("overall/<int:year>/<int:month>", MonthlyLeaderboard, name="MonthlyLeaderboard"),
     path("fullgame", FG_Leaderboard, name="FullGameLeaderboard"),
     path("player/<str:name>", PlayerProfile, name="PlayerProfile"),
+    path("player/<str:name>/history", PlayerHistory, name="PlayerHistory"),
     path("<str:abbr>/", GameLeaderboard, name="CategorySelection"),
     path("<str:game_abbr>/all", IL_Leaderboard, name="GameLeaderboard"),
     path("<str:abbr>/ils", ILGameLeaderboard, name="CategorySelection"),